[more strict warning setting for MSVC]
[Changes]
- if on Windows, the default toolchain is Visual Studio,
  and if multiple toolsets are installed, the following
  can be done:
gmake ROUND_MODE=UP CMAKE_OPTIONS="-T ClangCL"
gmake ROUND_MODE=UP CMAKE_OPTIONS="-T v141"
gmake ROUND_MODE=UP CMAKE_OPTIONS="-T v142"

+++ lib/src/kernels/pooling/mli_krn_pool_hwc.h
@@ -318,7 +318,7 @@ static MLI_FORCE_INLINE void mli_krn_pool_hwc_pad(
                     area.clmn_beg = 0;
                     area.clmn_end = out.width - CEIL_DIV(padding_right, stride_width);
                     // compensate for cases where kernel size is larger than input size
-                    area.clmn_end = MAX(CEIL_DIV(padding_left, stride_width), area.clmn_end);
+                    area.clmn_end = MAX(CEIL_DIV(padding_left, stride_width), (int)area.clmn_end);
                 } else {
                     continue;
                 }
